backend/pyproject.toml
[tool.poetry]
name = "munro"
version = "2.0.0"
description = "League Results. Sorted."
keywords = ['sports', 'leagues', 'scoring', 'results', 'orienteering']
authors = ["Ben Brown <9870007+brownben@users.noreply.github.com>"]
homepage = 'https://munroleagues.com'
repository = 'https://github.com/brownben/munro'
readme = 'README.md'
license = "MIT"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "^3.9"
asyncpg = "^0.27.0"
defusedxml = "^0.7.1"
fastapi = "^0.87.0"
google-auth = "^2.10.0"
httpx = "^0.23.0"
piccolo = "^0.97.0"
uvicorn = "^0.20.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
aiosqlite = "^0.17.0"
black = "^22.3.0"
coverage = "^6.3.2"
isort = "^5.10.1"
mypy = "^0.991"
respx = "^0.20.1"
time-machine = "^2.6.0"
types-requests = "^2.27.16"


[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
omit = [".venv/*", "*/test*", "*/*database_conf.py"]
exclude_lines = ["coverage: ignore"]

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
plugins = "pydantic.mypy"
mypy_path = 'src/types'
files = '**/*.py'
explicit_package_bases = true

[tool.ruff]
ignore = ["E501"]
per-file-ignores = { "./src/database/*"=["E712"] }
